Namespace juggling.
